+++ test/test_mongos_ha.py
@@ -12,15 +12,18 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Test MongoClient's mongos high-availability features using a mock."""
+"""Test MongoClient's mongos load balancing using a mock."""
 
 import sys
 import threading
+import time
 
 sys.path[0:0] = [""]
 
-from pymongo.errors import AutoReconnect
-from test import unittest, client_context, SkipTest, MockClientTest
+from pymongo.errors import AutoReconnect, InvalidOperation
+from pymongo.server_selectors import writable_server_selector
+from pymongo.topology_description import TOPOLOGY_TYPE
+from test import unittest, client_context, MockClientTest
 from test.pymongo_mocks import MockClient
 from test.utils import connected, wait_until
 
@@ -54,27 +57,39 @@ def do_simple_op(client, nthreads):
         assert t.passed
 
 
-class TestMongosHA(MockClientTest):
+def writable_addresses(topology):
+    return set(server.description.address for server in
+               topology.select_servers(writable_server_selector))
 
-    def mock_client(self):
-        return MockClient(
+
+class TestMongosLoadBalancing(MockClientTest):
+
+    def mock_client(self, **kwargs):
+        mock_client = MockClient(
             standalones=[],
             members=[],
             mongoses=['a:1', 'b:2', 'c:3'],
             host='a:1,b:2,c:3',
-            connect=False)
-        
-    def test_lazy_connect(self):
-        # TODO: Reimplement Mongos HA with PyMongo 3's MongoClient.
-        raise SkipTest('Mongos HA must be reimplemented with 3.0 MongoClient')
+            connect=False,
+            **kwargs)
 
+        # Latencies in seconds.
+        mock_client.mock_rtts['a:1'] = 0.020
+        mock_client.mock_rtts['b:2'] = 0.025
+        mock_client.mock_rtts['c:3'] = 0.040
+        return mock_client
+
+    def test_lazy_connect(self):
+        # While connected() ensures we can trigger connection from the main
+        # thread and wait for the monitors, this test triggers connection from
+        # several threads at once to check for data races.
         nthreads = 10
         client = self.mock_client()
         self.assertEqual(0, len(client.nodes))
 
         # Trigger initial connection.
         do_simple_op(client, nthreads)
-        self.assertEqual(3, len(client.nodes))
+        wait_until(lambda: len(client.nodes) == 3, 'connect to all mongoses')
 
     def test_reconnect(self):
         nthreads = 10
@@ -92,35 +107,21 @@ class TestMongosHA(MockClientTest):
                    'reconnect to all mongoses')
 
     def test_failover(self):
-        # TODO: PyMongo 3's MongoClient currently picks a new Mongos at random
-        #       for each operation (besides getMore). Need to "pin".
-        raise SkipTest('Mongos HA must be reimplemented with 3.0 MongoClient')
-
-        nthreads = 1
-
-        # ['1:1', '2:2', '3:3', ...]
-        mock_hosts = ['%d:%d' % (i, i) for i in range(50)]
-        client = MockClient(
-            standalones=[],
-            members=[],
-            mongoses=mock_hosts,
-            host=','.join(mock_hosts))
-
-        self.assertEqual(len(mock_hosts), len(client.nodes))
+        nthreads = 10
+        client = connected(self.mock_client(localThresholdMS=0.001))
+        wait_until(lambda: len(client.nodes) == 3, 'connect to all mongoses')
 
         # Our chosen mongos goes down.
-        client.kill_host('%s:%s' % (client.host, client.port))
+        client.kill_host('a:1')
 
-        # Trigger failover. AutoReconnect should be raised exactly once.
-        errors = []
+        # Trigger failover to higher-latency nodes. AutoReconnect should be
+        # raised at most once in each thread.
         passed = []
 
         def f():
             try:
                 client.db.command('ismaster')
             except AutoReconnect:
-                errors.append(True)
-
                 # Second attempt succeeds.
                 client.db.command('ismaster')
 
@@ -133,11 +134,45 @@ class TestMongosHA(MockClientTest):
         for t in threads:
             t.join()
 
-        self.assertEqual(1, len(errors))
         self.assertEqual(nthreads, len(passed))
 
-        # Down host is still in list.
-        self.assertEqual(len(mock_hosts), len(client.nodes))
+        # Down host removed from list.
+        self.assertEqual(2, len(client.nodes))
+
+    def test_local_threshold(self):
+        client = connected(self.mock_client(localThresholdMS=30))
+        wait_until(lambda: len(client.nodes) == 3, 'connect to all mongoses')
+        topology = client._topology
+
+        # All are within a 30-ms latency window, see self.mock_client().
+        self.assertEqual(set([('a', 1), ('b', 2), ('c', 3)]),
+                         writable_addresses(topology))
+
+    def test_load_balancing(self):
+        # Although the server selection JSON tests already prove that
+        # select_servers works for sharded topologies, here we do an end-to-end
+        # test of discovering servers' round trip times and configuring
+        # localThresholdMS.
+        client = connected(self.mock_client())
+        wait_until(lambda: len(client.nodes) == 3, 'connect to all mongoses')
+
+        # Prohibited for topology type Sharded.
+        with self.assertRaises(InvalidOperation):
+            client.address
+
+        topology = client._topology
+        self.assertEqual(TOPOLOGY_TYPE.Sharded,
+                         topology.description.topology_type)
+
+        # a and b are within the 15-ms latency window, see self.mock_client().
+        self.assertEqual(set([('a', 1), ('b', 2)]),
+                         writable_addresses(topology))
+
+        client.mock_rtts['a:1'] = 0.040
+
+        # Discover only b is within latency window.
+        wait_until(lambda: set([('b', 2)]) == writable_addresses(topology),
+                   'discover server "a" is too far')
 
 
 if __name__ == "__main__":